requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx
texsoup
sqlalchemy>=2.0.0
asyncpg
psycopg2-binary
slowapi
livekit
livekit-agents[google, turn-detector]
livekit-api
cryptography
aiohttp>=3.8.0
async-lru>=2.0.0
PyJWT[cryptography]>=2.8.0